import of sfio2005 from http://www2.research.att.com/sw/download/

+++ src/lib/sfio/Stdio_b/doscan.c
@@ -20,7 +20,7 @@ va_list args;	/* arg list if argf == 0 */
 	if(!(sf = _sfstream(f)))
 		return -1;
 
-	if((rv = sfvscanf(sf,form,args)) <= 0)
+	if((rv = (int)sfvscanf(sf,form,args)) <= 0)
 		_stdseterr(f,sf);
 	return rv;
 }
