+++ src/lib/sfio/Stdio_b/sscanf.c
@@ -23,7 +23,7 @@ va_dcl
 	form = va_arg(args,char*);
 #endif
 
-	rv = (s && form) ? sfvsscanf(s,form,args) : -1;
+	rv = (s && form) ? (int)sfvsscanf(s,form,args) : -1;
 	va_end(args);
 
 	return rv;
