+++ src/lib/sfio/Stdio_b/vfscanf.c
@@ -19,7 +19,7 @@ va_list args;           /* arg list if argf == 0 */
 	if(!form || !(sf = _sfstream(f)))
 		return -1;
 
-	if((rv = sfvscanf(sf,form,args)) <= 0)
+	if((rv = (int)sfvscanf(sf,form,args)) <= 0)
 		_stdseterr(f,sf);
 
 	return rv;
