+++ src/lib/sfio/Stdio_b/vprintf.c
@@ -18,7 +18,7 @@ va_list args;           /* arg list if argf == 0 */
 	if(!form || !(sf = _sfstream(stdout)))
 		return -1;
 
-	if((rv = sfvprintf(sf,form,args)) < 0)
+	if((rv = (int)sfvprintf(sf,form,args)) < 0)
 		_stdseterr(stdout,sf);
 
 	return rv;
