+++ src/lib/sfio/Stdio_b/vscanf.c
@@ -18,7 +18,7 @@ va_list args;           /* arg list if argf == 0 */
 	if(!form || !(sf = _sfstream(stdin)))
 		return -1;
 
-	if((rv = sfvscanf(sf,form,args)) <= 0)
+	if((rv = (int)sfvscanf(sf,form,args)) <= 0)
 		_stdseterr(stdin,sf);
 
 	return rv;
