+++ src/lib/sfio/Stdio_s/stdfscanf.c
@@ -1,5 +1,4 @@
-#include	"sfhdr.h"
-#include	"stdio.h"
+#include	"stdio_s.h"
 
 /*	Read formatted data from a stream
 **
@@ -26,7 +25,7 @@ va_dcl
 	form = va_arg(args,char*);
 #endif
 
-	rv = (f && form) ? sfvscanf(f,form,args) : -1;
+	rv = (f && form) ? (int)sfvscanf(f,form,args) : -1;
 
 	va_end(args);
 	return rv;
